requirements.txt
absl-py==2.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
filelock==3.14.0
fonttools==4.53.0
fsspec==2024.6.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.64.1
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
kiwisolver==1.4.5
lxml==5.2.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.82
packaging==24.1
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytils==0.4.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sympy==1.12.1
tbb==2021.12.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
tools==0.1.9
torch==2.3.1+cu118
torch-tb-profiler==0.4.3
torchaudio==2.3.1+cu118
torchvision==0.18.1+cu118
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.31
ultralytics-thop==0.2.8
urllib3==2.2.1
Werkzeug==3.0.3
